ai-services/requirements.txt
langchain
langchain-openai
python-dotenv
openai 